Fix height and weight conversions

+++ app/src/main/java/org/simple/clinic/patientattribute/entry/BMIEntryUiRenderer.kt
@@ -10,8 +10,8 @@ class BMIEntryUiRenderer(
   override fun render(model: BMIEntryModel) {
     if (model.height.isNotEmpty() && model.weight.isNotEmpty()) {
       val bmi = BMIReading(
-          height = model.height,
-          weight = model.weight
+          height = model.height.toFloat(),
+          weight = model.weight.toFloat()
       ).calculateBMI()
 
       ui.showBMI(bmi.toString())
